requirements.txt
telethon==1.21.1
python-dotenv==0.19.2
requests>=2.31.0
aiohttp>=3.9.1
asyncio>=3.4.3